+++ src/py/flwr/client/grpc_client/connection_test.py
@@ -25,7 +25,7 @@ from unittest.mock import patch
 import grpc
 
 from flwr.common import DEFAULT_TTL, ConfigsRecord, Message, Metadata, RecordDict
-from flwr.common import recordset_compat as compat
+from flwr.common import recorddict_compat as compat
 from flwr.common.constant import MessageTypeLegacy
 from flwr.common.retry_invoker import RetryInvoker, exponential
 from flwr.common.typing import Code, GetPropertiesRes, Status
