+++ src/py/flwr/client/mod/secure_aggregation/secaggplus_mod.py
@@ -30,7 +30,7 @@ from flwr.common import (
     ndarray_to_bytes,
     parameters_to_ndarrays,
 )
-from flwr.common import recordset_compat as compat
+from flwr.common import recorddict_compat as compat
 from flwr.common.constant import MessageType
 from flwr.common.logger import log
 from flwr.common.secure_aggregation.crypto.shamir import create_shares
