+++ src/py/flwr/server/superlink/fleet/vce/backend/raybackend_test.py
@@ -37,7 +37,7 @@ from flwr.common import (
     Scalar,
 )
 from flwr.common.constant import PARTITION_ID_KEY
-from flwr.common.recordset_compat import getpropertiesins_to_recordset
+from flwr.common.recorddict_compat import getpropertiesins_to_recordset
 from flwr.server.superlink.fleet.vce.backend.backend import BackendConfig
 from flwr.server.superlink.fleet.vce.backend.raybackend import RayBackend
