+++ src/py/flwr/server/superlink/serverappio/serverappio_servicer_test.py
@@ -366,7 +366,7 @@ class TestServerAppIoServicer(unittest.TestCase):  # pylint: disable=R0902
             run_id=run_id,
             node_id=0,
             node_config=maker.user_config(),
-            state=maker.recordset(1, 1, 1),
+            state=maker.recorddict(1, 1, 1),
             run_config=maker.user_config(),
         )
 
@@ -417,7 +417,7 @@ class TestServerAppIoServicer(unittest.TestCase):  # pylint: disable=R0902
             run_id=run_id,
             node_id=0,
             node_config=maker.user_config(),
-            state=maker.recordset(1, 1, 1),
+            state=maker.recorddict(1, 1, 1),
             run_config=maker.user_config(),
         )
